Provider tag interfaces done right
