+++ pipeline-braille-utils/common-utils/src/main/java/org/daisy/pipeline/braille/common/BrailleTranslator.java
@@ -6,4 +6,8 @@ package org.daisy.pipeline.braille.common;
  * The {@link #transform(String)} and {@link #transform(String[])} methods
  * return unicode braille.
  */
-public interface BrailleTranslator extends TextTransform {}
+public interface BrailleTranslator extends TextTransform {
+	
+	public interface Provider<T extends BrailleTranslator> extends TextTransform.Provider<T> {}
+	
+}
