+++ pipeline-braille-utils/common-utils/src/main/java/org/daisy/pipeline/braille/common/MathMLTransform.java
@@ -4,4 +4,8 @@ package org.daisy.pipeline.braille.common;
  * A {@link org.daisy.pipeline.braille.common.Transform} that can transform
  * a MathML node to text.
  */
-public interface MathMLTransform extends Transform {}
+public interface MathMLTransform extends Transform {
+	
+	public interface Provider<T extends MathMLTransform> extends Transform.Provider<T> {}
+	
+}
