+++ pipeline-braille-utils/common-utils/src/main/java/org/daisy/pipeline/braille/common/Transform.java
@@ -6,4 +6,8 @@ package org.daisy.pipeline.braille.common;
  * Classes that implement this interface must have some kind of "transform"
  * method.
  */
-public interface Transform {}
+public interface Transform {
+	
+	public interface Provider<T extends Transform> extends org.daisy.pipeline.braille.common.Provider<String,T> {}
+	
+}
