--- pipeline-braille-utils/common-utils/src/main/java/org/daisy/pipeline/braille/common/XProcCSSBlockTransform.java
@@ -1,7 +1,0 @@
-package org.daisy.pipeline.braille.common;
-
-public interface XProcCSSBlockTransform extends CSSBlockTransform, XProcTransform {
-	
-	public interface Provider extends XProcTransform.Provider<XProcCSSBlockTransform> {}
-	
-}
