--- pipeline-braille-utils/common-utils/src/main/java/org/daisy/pipeline/braille/common/XProcCSSStyledDocumentTransform.java
@@ -1,3 +1,0 @@
-package org.daisy.pipeline.braille.common;
-
-public interface XProcCSSStyledDocumentTransform extends CSSStyledDocumentTransform, XProcTransform {}
