+++ pipeline-braille-utils/libhyphen/libhyphen-core/pom.xml
@@ -102,6 +102,8 @@
             <Service-Component>
               OSGI-INF/libhyphen-table-registry.xml
             </Service-Component>
+            <Export-Package>org.daisy.pipeline.braille.libhyphen</Export-Package>
+            <Private-Package>org.daisy.pipeline.braille.libhyphen.impl</Private-Package>
             <_dsannotations>*</_dsannotations>
           </instructions>
         </configuration>
