+++ pipeline-braille-utils/libhyphen/libhyphen-core/src/main/java/org/daisy/pipeline/braille/libhyphen/LibhyphenHyphenator.java
@@ -16,4 +16,6 @@ public abstract class LibhyphenHyphenator implements Hyphenator {
 	
 	public abstract URI asLibhyphenTable();
 	
+	public interface Provider extends Hyphenator.Provider<LibhyphenHyphenator> {}
+	
 }
