--- pipeline-braille-utils/liblouis/liblouis-core/src/main/java/org/daisy/pipeline/braille/liblouis/Liblouis.java
@@ -1,5 +1,0 @@
-package org.daisy.pipeline.braille.liblouis;
-
-import org.daisy.pipeline.braille.common.TextTransform;
-
-public interface Liblouis extends TextTransform.Provider<LiblouisTranslator> {}
