+++ pipeline-braille-utils/liblouis/liblouis-core/src/main/java/org/daisy/pipeline/braille/liblouis/LiblouisTranslator.java
@@ -2,6 +2,7 @@ package org.daisy.pipeline.braille.liblouis;
 
 import java.net.URI;
 
+import org.daisy.pipeline.braille.common.TextTransform;
 import org.daisy.pipeline.braille.common.CSSStyledTextTransform;
 
 public abstract class LiblouisTranslator implements CSSStyledTextTransform { // BrailleTranslator, Hyphenator
@@ -40,5 +41,7 @@ public abstract class LiblouisTranslator implements CSSStyledTextTransform { // 
 	public abstract String display(String braille);
 	
 	public abstract URI[] asLiblouisTable();
+	
+	public interface Provider extends TextTransform.Provider<LiblouisTranslator> {}
 	
 }
